1.0.12

+++ src/main/java/dev/saperate/elementals/entities/water/WaterShieldEntity.java
@@ -1,6 +1,7 @@
 package dev.saperate.elementals.entities.water;
 
 import dev.saperate.elementals.data.Bender;
+import dev.saperate.elementals.entities.common.AbstractElementalsEntity;
 import net.minecraft.entity.*;
 import net.minecraft.entity.data.DataTracker;
 import net.minecraft.entity.data.TrackedData;
@@ -19,7 +20,7 @@ import static dev.saperate.elementals.entities.ElementalEntities.WATERSHIELD;
 import static dev.saperate.elementals.utils.SapsUtils.summonParticles;
 
 
-public class WaterShieldEntity extends Entity {
+public class WaterShieldEntity extends AbstractElementalsEntity {
     private static final TrackedData<Integer> OWNER_ID = DataTracker.registerData(WaterShieldEntity.class, TrackedDataHandlerRegistry.INTEGER);
 
 
@@ -37,12 +38,6 @@ public class WaterShieldEntity extends Entity {
         setOwner(owner);
     }
 
-
-    @Override
-    protected void initDataTracker() {
-        this.getDataTracker().startTracking(OWNER_ID, 0);
-    }
-
     @Override
     public void tick() {
         super.tick();
@@ -53,34 +48,11 @@ public class WaterShieldEntity extends Entity {
 
         PlayerEntity owner = getOwner();
 
-        if(owner == null){
-            discard();
+        if(owner == null || isRemoved()){
             return;
         }
 
-        setPos(owner.getX(),owner.getY(),owner.getZ());
-
-
-        List<LivingEntity> hits = getWorld().getEntitiesByClass(LivingEntity.class,
-                getWorld().isClient ? getBoundingBox() : getBoundingBox().offset(getPos()),
-                LivingEntity::isAlive);
-
-        for (LivingEntity entity : hits) {
-           if(entity != owner){
-               //Naughty entities can still somehow get inside the shield. This pushes them back out
-               Vec3d direction = entity.getPos().add(0,1.5f,0).subtract(getPos()).multiply(0.1f);
-               entity.setVelocity(getVelocity().add(direction));
-           }
-        }
-
-        List<ProjectileEntity> projectiles = getWorld().getEntitiesByClass(ProjectileEntity.class,
-                getWorld().isClient ? getBoundingBox().expand(1.5f) : getBoundingBox().expand(1.5f).offset(getPos()),
-                ProjectileEntity::isAlive);
-
-        for (ProjectileEntity e : projectiles){
-            Vec3d direction = e.getPos().add(0,1.7f,0).subtract(getPos()).multiply(0.1f);
-            e.setVelocity(getVelocity().add(direction));
-        }
+        moveEntityTowardsGoal(owner.getPos().toVector3f());
     }
 
     @Override
@@ -88,7 +60,7 @@ public class WaterShieldEntity extends Entity {
         summonParticles( this,random, ParticleTypes.SPLASH, 10,100);
         if(!this.getWorld().isClient){
             Bender bender = Bender.getBender(getOwner());
-            if(bender != null){//Clean up the mess
+            if(bender != null && bender.currAbility != null){//Clean up the mess
                 bender.abilityData = null;
                 bender.currAbility.onRemove(bender);
             }
@@ -102,27 +74,17 @@ public class WaterShieldEntity extends Entity {
     }
 
     @Override
-    public boolean canHit() {
+    public boolean teleportsToGoal() {
         return true;
     }
 
     @Override
-    protected void readCustomDataFromNbt(NbtCompound nbt) {
-
+    public float projectileDeflectionRange() {
+        return 1.5f;
     }
 
     @Override
-    protected void writeCustomDataToNbt(NbtCompound nbt) {
-
-    }
-
-
-    public PlayerEntity getOwner() {
-        Entity owner = this.getWorld().getEntityById(this.getDataTracker().get(OWNER_ID));
-        return (owner instanceof LivingEntity) ? (PlayerEntity) owner : null;
-    }
-
-    public void setOwner(LivingEntity owner) {
-        this.getDataTracker().set(OWNER_ID, owner.getId());
+    public boolean discardsOnNullOwner() {
+        return true;
     }
 }
